# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DestinationContextRequest": {
        "additionalProperties": false,
        "properties": {
          "bias_radius_meters": {
            "anyOf": [
              {
                "maximum": 50000.0,
                "minimum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bias Radius Meters"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchCoordinateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "route_bounds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DestinationBoundsRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "DestinationContextRequest",
        "type": "object"
      },
      "DestinationResolveRequest": {
        "additionalProperties": false,
        "properties": {
          "provider": {
            "const": "google_places_new",
            "title": "Provider",
            "type": "string"
          },
          "provider_ref": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Provider Ref",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "session_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "revision",
          "provider",
          "provider_ref"
        ],
        "title": "DestinationResolveRequest",
        "type": "object"
      },
      "DestinationResolveResponse": {
        "additionalProperties": false,
        "properties": {
          "navigation_target": {
            "$ref": "#/components/schemas/compass__api__search__NavigationTargetResponse"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "selection": {
            "$ref": "#/components/schemas/compass__api__search__ResolvedSelectionResponse"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "revision",
          "selection",
          "navigation_target"
        ],
        "title": "DestinationResolveResponse",
        "type": "object"
      },
...
      "DestinationSuggestRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DestinationContextRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "intent": {
            "enum": [
              "ORIGIN_SEARCH",
              "DESTINATION_SEARCH"
            ],
            "title": "Intent",
            "type": "string"
          },
          "language": {
            "default": "it",
            "pattern": "^[a-z]{2}$",
            "title": "Language",
            "type": "string"
          },
          "operation": {
            "default": "autocomplete",
            "enum": [
              "autocomplete",
              "text_search"
            ],
            "title": "Operation",
            "type": "string"
          },
          "query": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "session_id": {
            "maxLength": 36,
            "minLength": 36,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "intent",
          "query",
          "session_id",
          "revision"
        ],
        "title": "DestinationSuggestRequest",
        "type": "object"
      },
      "DestinationSuggestResponse": {
        "additionalProperties": false,
        "properties": {
          "maximum_results": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Maximum Results",
            "type": "integer"
          },
          "provider": {
            "const": "google_places_new",
            "default": "google_places_new",
            "title": "Provider",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DestinationSuggestionResponse"
            },
            "title": "Results",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "revision",
          "results"
        ],
        "title": "DestinationSuggestResponse",
        "type": "object"
      },
      "DestinationSuggestionResponse": {
        "additionalProperties": false,
        "properties": {
          "address_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Preview"
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "distance_meters": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Meters"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "business",
              "address",
              "locality",
              "unknown"
            ],
            "title": "Kind",
            "type": "string"
          },
          "marginal_added_duration_seconds": {
            "title": "Marginal Added Duration Seconds",
            "type": "null"
          },
          "provider": {
            "const": "google_places_new",
            "title": "Provider",
            "type": "string"
          },
          "provider_rank": {
            "minimum": 0.0,
            "title": "Provider Rank",
            "type": "integer"
          },
          "provider_ref": {
            "title": "Provider Ref",
            "type": "string"
          },
          "requires_resolution": {
            "title": "Requires Resolution",
            "type": "boolean"
          },
          "search_intent": {
            "title": "Search Intent",
            "type": "null"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_added_duration_seconds": {
            "title": "Total Added Duration Seconds",
            "type": "null"
          },
          "within_time_budget": {
            "title": "Within Time Budget",
            "type": "null"
          }
        },
        "required": [
          "id",
          "provider",
          "provider_ref",
          "kind",
          "title",
          "subtitle",
          "address_preview",
          "provider_rank",
          "requires_resolution",
          "attribution"
        ],
        "title": "DestinationSuggestionResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlaceSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "cacheable": {
            "title": "Cacheable",
            "type": "boolean"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/PlaceSearchResultResponse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "cacheable",
          "results"
        ],
        "title": "PlaceSearchResponse",
        "type": "object"
      },
...
      "compass__api__search__NavigationTargetResponse": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "$ref": "#/components/schemas/SearchCoordinateResponse"
          },
          "map_label": {
            "const": "Destinazione selezionata",
            "title": "Map Label",
            "type": "string"
          },
          "provider_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Ref"
          }
        },
        "required": [
          "location",
          "provider_ref",
          "map_label"
        ],
        "title": "NavigationTargetResponse",
        "type": "object"
      },
      "compass__api__search__ResolvedSelectionResponse": {
        "additionalProperties": false,
        "properties": {
          "address_components": {
            "items": {
              "$ref": "#/components/schemas/AddressComponentResponse"
            },
            "title": "Address Components",
            "type": "array"
          },
          "attribution": {
            "items": {
              "type": "string"
            },
            "title": "Attribution",
            "type": "array"
          },
          "field_sources": {
            "additionalProperties": {
              "const": "google_places_new",
              "type": "string"
            },
            "title": "Field Sources",
            "type": "object"
          },
          "formatted_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formatted Address"
          },
          "kind": {
            "enum": [
              "business",
              "address",
              "locality",
              "unknown"
            ],
            "title": "Kind",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/SearchCoordinateResponse"
          },
          "normalized_address": {
            "$ref": "#/components/schemas/NormalizedAddressResponse"
          },
          "provider": {
            "const": "google_places_new",
            "title": "Provider",
            "type": "string"
          },
          "provider_ref": {
            "title": "Provider Ref",
            "type": "string"
          },
          "resolution_status": {
            "const": "resolved",
            "title": "Resolution Status",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "provider_ref",
          "formatted_address",
          "address_components",
          "normalized_address",
          "location",
          "kind",
          "resolution_status",
          "attribution",
          "field_sources"
        ],
        "title": "ResolvedSelectionResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OptionalHTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CNG-aware navigation API foundations for Italy.",
    "title": "Compass CNG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/destinations/resolve": {
      "post": {
        "operationId": "destination_resolve_api_v1_destinations_resolve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationResolveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Destination Resolve",
        "tags": [
          "place-search"
        ]
      }
    },
    "/api/v1/destinations/suggest": {
      "post": {
        "operationId": "destination_suggest_api_v1_destinations_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DestinationSuggestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationSuggestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Destination Suggest",
        "tags": [
          "place-search"
        ]
      }
    },
    "/api/v1/places/search": {
      "get": {
        "operationId": "place_search_api_v1_places_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 20,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "language",
            "required": false,
            "schema": {
              "default": "it",
              "pattern": "^[A-Za-z]{2}(-[A-Za-z]{2})?$",
              "title": "Language",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaceSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid Compass API credentials."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid search-provider response."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Search provider unavailable."
          }
        },
        "security": [
          {
            "OptionalHTTPBasic": []
          }
        ],
        "summary": "Place Search",
        "tags": [
          "place-search"
        ]
      }
    },
...
  }
}